requirements.txt
# --- For-Bot-Working --------- #
pyrogram
TgCrypto
pyromod==1.5
# --- For-Database ------------ #
pymongo
dnspython
python-dotenv